Fix warnings (ITK, filterPreview)

+++ modules/FilterPreview/iAFilterPreviewModuleInterface.cpp
@@ -267,7 +267,7 @@ void iAFilterPreviewModuleInterface::openSplitView(iASlicerImpl* slicer, const Q
 
 				if (!slicerCopies.empty())
 				{
-					for (int k = 0; k < slicerCopies.size(); ++k)
+					for (size_t k = 0; k < slicerCopies.size(); ++k)
 					{
 						iASlicerImpl* slicerCopy = slicerCopies[k];  // Access the slicer copy by index
 
@@ -362,7 +362,7 @@ void iAFilterPreviewModuleInterface::openSplitView(iASlicerImpl* slicer, const Q
 			if (!slicerCopies.empty())
 			{
 
-				for (int i = 0; i < slicerCopies.size(); ++i)
+				for (size_t i = 0; i < slicerCopies.size(); ++i)
 				{
 					iASlicerImpl* slicerCopy = slicerCopies[i];  // Access the slicer copy by index
